Try: Block Library polish
This starts work on https://github.com/WordPress/gutenberg/issues/17335#issuecomment-562049752.

For starters, it redesigns the block library:

- It is full-height to give lots more space.
- It drops the help panel, as we now have a welcome screen.
- It moves the tip to the footer.
- It detaches the preview and shows it only on hover.

Still to do here, is to polish the responsive behavior, possibly apply the new category names, and massage the pixels a bit further.

What should also be done is to explore the UI from #18667, as well as add a tab for block patterns, and finally the ability for the block library to open as a modal dialog defaulting to the block patterns tab when opened from the sibling inserter. However these are probably best explored separately.